LayoutTests/inspector/script-profiler/tracking-expected.txt

Tests that ScriptProfiler.startTracking and ScriptProfiler.stopTracking trigger trackingStart and trackingComplete events.


== Running test suite: ScriptProfiler.Tracking
-- Running test case: ScriptProfiler.Tracking.StartAndStopTracking
ScriptProfiler.trackingStart
PASS: Should have a timestamp when starting.
ScriptProfiler.trackingComplete
